internal/templates/conversational/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
redis==5.0.0
psycopg2-binary==2.9.9
openai==1.63.0
sse-starlette==2.2.1
